# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Script to Image Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Image orientation preference
 */
export type ImageSearchInputOrientation = typeof ImageSearchInputOrientation[keyof typeof ImageSearchInputOrientation];


export const ImageSearchInputOrientation = {
  landscape: 'landscape',
  portrait: 'portrait',
  square: 'square',
} as const;

export interface ImageSearchInput {
  /** The search query derived from the script line */
  query: string;
  /** The line number in the script */
  lineNumber: number;
  /** The original script line text */
  lineText: string;
  /** Image provider to use (pexels, unsplash) */
  provider?: string;
  /**
     * Number of images to fetch
     * @minimum 1
     * @maximum 8
     */
  perPage?: number;
  /** Image orientation preference */
  orientation?: ImageSearchInputOrientation;
  /** Whether to enable safe search */
  safeSearch?: boolean;
}

export interface ImageResult {
  id: string;
  /** Full resolution image URL */
  url: string;
  /** Thumbnail URL for display */
  thumbnailUrl: string;
  /** Medium resolution URL */
  mediumUrl?: string;
  photographer: string;
  photographerUrl: string;
  /** Provider name (pexels, unsplash) */
  source: string;
  width: number;
  height: number;
  /**
     * Alt text for the image
     * @nullable
     */
  alt?: string | null;
}

export interface ImageSearchResult {
  lineNumber: number;
  lineText: string;
  /** The actual query used for the search */
  query: string;
  images: ImageResult[];
  provider: string;
  totalResults?: number;
}

export interface ImageSettings {
  availableProviders: string[];
  defaultProvider: string;
  maxPerPage?: number;
}