# lib/api-zod/src/generated/types/imageResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Script to Image Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface ImageResult {
  id: string;
  /** Full resolution image URL */
  url: string;
  /** Thumbnail URL for display */
  thumbnailUrl: string;
  /** Medium resolution URL */
  mediumUrl?: string;
  photographer: string;
  photographerUrl: string;
  /** Provider name (pexels, unsplash) */
  source: string;
  width: number;
  height: number;
  /**
     * Alt text for the image
     * @nullable
     */
  alt?: string | null;
}