# lib/api-zod/src/generated/types/imageSearchInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Script to Image Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { ImageSearchInputOrientation } from './imageSearchInputOrientation';

export interface ImageSearchInput {
  /** The search query derived from the script line */
  query: string;
  /** The line number in the script */
  lineNumber: number;
  /** The original script line text */
  lineText: string;
  /** Image provider to use (pexels, unsplash) */
  provider?: string;
  /**
     * Number of images to fetch
     * @minimum 1
     * @maximum 8
     */
  perPage?: number;
  /** Image orientation preference */
  orientation?: ImageSearchInputOrientation;
  /** Whether to enable safe search */
  safeSearch?: boolean;
}