# lib/api-zod/src/generated/types/imageSearchResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Script to Image Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { ImageResult } from './imageResult';

export interface ImageSearchResult {
  lineNumber: number;
  lineText: string;
  /** The actual query used for the search */
  query: string;
  images: ImageResult[];
  provider: string;
  totalResults?: number;
}