# lib/api-zod/src/generated/types/imageSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Script to Image Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface ImageSettings {
  availableProviders: string[];
  defaultProvider: string;
  maxPerPage?: number;
}